v1.1.0: rename to ruby-gem-setup-github-packages-action
The previous naming was confusing and contentious. The GPR acronym is
oddly hard to remember, not that widly used and I'm even unsure if the
R is registry or repository! Fix this by just calling it github-packages
and not using any abbreviations.

Additionally rubygems- made it look like it was for the actual rubygems
site and registry. The intention was to show this is a ruby action
dealing with gems. Changing the prefix to `ruby-gem-` to fix that.

Note this also updates the name for ruby-gem-push-action in the docs,
as that is due to be renamed to match this convention.